core/src/requirements.txt
absl-py==1.4.0
annotated-types==0.6.0
array-record==0.5.0
astroid==3.0.1
astunparse==1.6.3
autopep8==2.0.4
beautifulsoup4==4.12.2
black==23.11.0
cachetools==5.3.2
Cerberus==1.3.5
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
dill==0.3.7
distlib==0.3.7
dm-tree==0.1.8
docopt==0.6.2
etils==1.5.2
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.46.0
fsspec==2023.12.1
gast==0.5.4
gcpds-image-segmentation @ git+https://github.com/UN-GCPDS/python-gcpds.image_segmentation.git@2263536d47a757e71ca0bc62127a1ec22943d762
gdown==4.7.1
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
h5py==3.10.0
idna==3.6
imageio==2.33.0
importlib-resources==6.1.1
iniconfig==2.0.0
isort==5.13.0
joblib==1.3.2
keras==2.15.0
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
mccabe==0.7.0
ml-dtypes==0.2.0
mypy==1.7.1
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.2
nvidia-cublas-cu12==12.2.5.6
nvidia-cuda-cupti-cu12==12.2.142
nvidia-cuda-nvcc-cu12==12.2.140
nvidia-cuda-nvrtc-cu12==12.2.140
nvidia-cuda-runtime-cu12==12.2.140
nvidia-cudnn-cu12==8.9.4.25
nvidia-cufft-cu12==11.0.8.103
nvidia-curand-cu12==10.3.3.141
nvidia-cusolver-cu12==11.5.2.141
nvidia-cusparse-cu12==12.1.2.141
nvidia-nccl-cu12==2.16.5
nvidia-nvjitlink-cu12==12.2.140
oauthlib==3.2.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
pathspec==0.12.0
pep517==0.13.1
Pillow==10.1.0
pip-api==0.0.30
pipreqs==0.4.13
platformdirs==4.1.0
plette==0.4.4
pluggy==1.3.0
promise==2.3
protobuf==3.20.3
psutil==5.9.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pydantic==2.5.2
pydantic_core==2.14.5
pylint==3.0.2
pyparsing==3.1.1
PySocks==1.7.1
pytest==7.4.3
python-dateutil==2.8.2
requests==2.31.0
requests-oauthlib==1.3.1
requirementslib==3.0.0
rsa==4.9
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
soupsieve==2.5
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0.post1
tensorflow-datasets==4.9.3
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-metadata==1.14.0
termcolor==2.4.0
threadpoolctl==3.2.0
tifffile==2023.12.9
toml==0.10.2
tomlkit==0.12.3
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
wrapt==1.14.1
yarg==0.1.9
zipp==3.17.0
